Add components and functionality for the fillout-th (#1)
This is Ron's submission to the alternative navigation menu for Fill Out.

Draggability to the buttons
Clicking on the buttons opens a pop up
Add button in between the draggable buttons
Add button at the end
If there are too many buttons, it has the ability of mousing over horizontally to view the hidden buttons.